Clean up and fix formatting
Move Four Way shape logic to Abstract tape class
Revert name of abstract falling block class to AbstractBlockMinecoloniesFalling
General formatting and doc fixes
