+++ src/api/java/com/minecolonies/api/blocks/AbstractBlockMinecoloniesFalling.java
@@ -1,0 +1,40 @@
+package com.minecolonies.api.blocks;
+
+import com.minecolonies.api.blocks.interfaces.IBlockMinecolonies;
+import com.minecolonies.api.util.constant.Suppression;
+import net.minecraft.block.*;
+import net.minecraft.item.BlockItem;
+import net.minecraft.item.Item;
+import net.minecraftforge.registries.IForgeRegistry;
+
+public abstract class AbstractBlockMinecoloniesFalling<B extends AbstractBlockMinecoloniesFalling<B>> extends FallingBlock implements IBlockMinecolonies<B>
+{
+    public AbstractBlockMinecoloniesFalling(final Properties properties)
+    {
+        super(properties);
+    }
+
+    /**
+     * Registery block at gameregistry.
+     *
+     * @param registry the registry to use.
+     * @return the block itself.
+     */
+    @Override
+    @SuppressWarnings(Suppression.UNCHECKED)
+    public B registerBlock(final IForgeRegistry<Block> registry)
+    {
+        registry.register(this);
+        return (B) this;
+    }
+    /**
+     * Registery block at gameregistry.
+     *
+     * @param registry the registry to use.
+     */
+    @Override
+    public void registerBlockItem(final IForgeRegistry<Item> registry, final Item.Properties properties)
+    {
+        registry.register((new BlockItem(this, properties)).setRegistryName(this.getRegistryName()));
+    }
+}
